+++ data_control.cpp
@@ -1,0 +1,38 @@
+
+#include "data_control.h"
+
+namespace unit_control
+{
+    namespace flags
+    {
+        bool is_behind(CUnit* unit) {  return unit->Flags & UNIT_FLAG_BEHIND;  }
+        bool is_guard(CUnit* unit) {  return unit->Flags & UNIT_FLAG_GUARDING;  }
+        bool is_hold(CUnit* unit) {  return unit->Flags & UNIT_FLAG_HOLDING;  }
+        bool is_noaid(CUnit* unit) {  return unit->Flags & UNIT_FLAG_RECEIVING_NO_AID;  }
+        bool is_avoid(CUnit* unit) {  return unit->Flags & UNIT_FLAG_AVOIDING;  }
+        bool is_nocross(CUnit* unit) {  return unit->Flags & UNIT_FLAG_NO_CROSS_WATER;  }
+        bool is_reveal(CUnit* unit, std::string flag) {
+            for(size_t i = 0; i < flag.size(); ++i)
+                flag[i] = std::tolower(flag[i]);
+            if (!flag.compare("unit"))
+                return unit->Flags & UNIT_FLAG_REVEALING_UNIT;
+            if (!flag.compare("faction"))
+                return unit->Flags & UNIT_FLAG_REVEALING_FACTION;
+            return false;
+        }
+        bool is_spoils(CUnit* unit, const std::string flag) {
+            // not implemented
+            return false;
+        }
+
+        bool is_consume(CUnit* unit, std::string flag) {
+            for(size_t i = 0; i < flag.size(); ++i)
+                flag[i] = std::tolower(flag[i]);
+            if (!flag.compare("unit"))
+                return unit->Flags & UNIT_FLAG_CONSUMING_UNIT;
+            if (!flag.compare("faction"))
+                return unit->Flags & UNIT_FLAG_CONSUMING_FACTION;
+            return false;
+        }
+    }
+}
