+++ data_control.h
@@ -1,0 +1,37 @@
+#ifndef DATA_CONTROL_INCL
+#define DATA_CONTROL_INCL
+
+#include "data.h"
+
+namespace unit_control
+{
+    namespace flags
+    {
+        bool is_behind(CUnit* unit);
+        bool is_guard(CUnit* unit);
+        bool is_hold(CUnit* unit);
+        bool is_noaid(CUnit* unit);
+        bool is_avoid(CUnit* unit);
+        bool is_nocross(CUnit* unit);
+        bool is_reveal(CUnit* unit, std::string flag);
+        bool is_spoils(CUnit* unit, std::string flag);
+        bool is_consume(CUnit* unit, std::string flag);
+        bool is_leader(CUnit* unit);
+    }
+}
+
+namespace land_control
+{
+    template<typename T>
+    void get_units_if(CLand* land, std::vector<CUnit*>& units, T Pred)
+    {
+        for (size_t i = 0; i < land->Units.Count(); i++)
+        {
+            CUnit* unit = (CUnit*)land->Units.At(i);
+            if (Pred(unit))
+                units.push_back(unit);
+        }       
+    }
+}
+
+#endif
