+++ unitsplitdlg.h
@@ -45,130 +45,4 @@ private:
     DECLARE_EVENT_TABLE()
 };
 
-#include "data.h"
-#include <vector>
-#include <unordered_map>
-#include <string>
-#include "wx/checkbox.h"
-#include "wx/spinctrl.h"
-//#include "wx/wx.h"
-
-namespace unit_control
-{
-    namespace flags
-    {
-        bool is_behind(CUnit* unit);
-        bool is_guard(CUnit* unit);
-        bool is_hold(CUnit* unit);
-        bool is_noaid(CUnit* unit);
-        bool is_avoid(CUnit* unit);
-        bool is_nocross(CUnit* unit);
-        bool is_reveal(CUnit* unit, std::string flag);
-        bool is_spoils(CUnit* unit, std::string flag);
-        bool is_consume(CUnit* unit, std::string flag);
-    }
-};
-
-namespace land_control
-{
-    template<typename T>
-    void get_units_if(CLand* land, std::vector<CUnit*>& units, T Pred)
-    {
-        for (size_t i = 0; i < land->Units.Count(); i++)
-        {
-            CUnit* unit = (CUnit*)land->Units.At(i);
-            if (Pred(unit))
-                units.push_back(unit);
-        }       
-    }
-};
-
-namespace game_control
-{
-    struct Skill
-    {
-        long study_price_;
-        std::string short_name_;
-        std::string long_name_;
-    };
-
-    const std::vector<Skill> get_skills();
-}
-
-
-
-class CCreateNewUnit : public CResizableDlg
-{
-    //1. studying should be multiplied by number of bought units
-    //1.2 and sorted out
-    //1.3 and jump to skill by letter
-    //2. need to remove duplicates in list of skills and realize why do we have them
-    //3. whitespaces at give order
-    //4. give_all_need button.
-    //4.2 units to give silver should be sorted out
-    //5. when multiply units also give should be multiplied
-
-
-    CUnit* unit_;
-    CLand* land_;
-    std::unordered_map<std::string, CProductMarket> sale_products_;
-    std::map<std::string, long> silver_holders_;
-
-    //naming section
-    wxSpinCtrl* spin_new_num_alias_;
-    wxTextCtrl* text_name_;
-    wxTextCtrl* text_loc_description_;
-    wxTextCtrl* text_description_;
-
-    //buy section
-    wxSpinCtrl* spin_buy_units_amount_;
-    wxComboBox* combobox_buy_units_type_;
-    wxCheckBox* flag_buy_repeating_;
-    wxCheckBox* flag_buy_all_;
-
-    //study section
-    wxCheckBox* flag_check_study_;
-    wxComboBox* combobox_skills_;
-    wxCheckBox* flag_study_repeating_;
-
-    //receive silver section
-    wxSpinCtrl* spin_silver_amount_;
-    wxComboBox* combobox_units_;
-    wxCheckBox* flag_receive_silver_repeating_;
-
-    //Expenses section
-    wxStaticText* expenses_buying_; 
-    wxStaticText* expenses_studying_; 
-    wxStaticText* expenses_all_; 
-
-    //flags section
-    wxCheckBox* flag_behind_;
-    wxCheckBox* flag_avoid_;
-    wxCheckBox* flag_hold_;
-    wxCheckBox* flag_noaid_;
-    wxCheckBox* flag_guard_;
-    wxCheckBox* flag_nocross_;
-
-    wxRadioBox* radiobox_flag_reveal_;
-    wxRadioBox* radiobox_flag_spoils_;
-    wxRadioBox* radiobox_flag_consume_;
-
-    //Copies
-    wxSpinCtrl* spin_copies_amount_;
-
-public:
-    //flags consist of command and it's possible values
-    CCreateNewUnit(wxWindow *parent, CUnit * pUnit, CLand* pLand);
-    ~CCreateNewUnit();
-
-private:
-    void UpdateExpences();
-    void OnCancel       (wxCommandEvent& event);
-    void OnOk           (wxCommandEvent& event);
-    void onAnySpinUpdate(wxSpinEvent& event);
-    void onAnyComboBoxUpdate(wxCommandEvent& event);
-
-    DECLARE_EVENT_TABLE()
-};
-
 #endif
